Remove "canonicalization" of OF rules matches for OF1.2
Some Accton based switches treat exact match as masked match with
zero(FULL_MASK) mask. This lead to false positive matches.

As workaround for this issue masked match with ZERO_MASK is used... But
canonicalization process drops mask for such matches.

So to be able to use mentioned workaround, disabling
canonicalization(trivial masks (all-0, all-1) are converted to the main
representation) for OF1.2.